+++ DirectX 11 Playground/Src/Scenegraph And Physics/Ray.cpp
@@ -127,16 +127,16 @@ void Ray::draw(ID3D11Device* device, ID3D11DeviceContext * context, VertexConsta
 	{
 		hasInit = true;
 
-		HRESULT hr = m_vertexBuffer.init(device, obbVertices.data(), obbVertices.size());
+		HRESULT hr = m_vertexBuffer.init(device, obbVertices.data(), static_cast<UINT>(obbVertices.size()));
 		COM_ERROR_IF_FAILED(hr, "Failed to initialize dynamic vertex buffer for Ray");
-		hr = m_indexBuffer.init(device, obbIndices.data(), obbIndices.size());
+		hr = m_indexBuffer.init(device, obbIndices.data(), static_cast<UINT>(obbIndices.size()));
 		COM_ERROR_IF_FAILED(hr, "Failed to initialize indices buffer for Ray");
 	}
 	else
 	{
 		if (isDirty)
 		{
-			m_vertexBuffer.updateBuffer(context, obbVertices.data(), obbVertices.size());
+			m_vertexBuffer.updateBuffer(context, obbVertices.data(), static_cast<UINT>(obbVertices.size()));
 			isDirty = false;
 		}
